added auth and added access controll to the delete route

+++ internal/server/CRUD.go
@@ -1,6 +1,7 @@
 package server
 
 import (
+	"database/sql"
 	"errors"
 	"fmt"
 	"log"
@@ -213,7 +214,42 @@ func (s *Server) UpdateRecipe(c *gin.Context) {
 }
 
 func (s *Server) DeleteRecipe(c *gin.Context) {
+	user, _ := c.Get("user")
+	if user == nil {
+		error_handler.HandleError(c, http.StatusUnauthorized, "User not logged in", []error{errors.New("user not logged in")})
+		c.AbortWithStatus(http.StatusUnauthorized)
+		return
+	}
+	userModel, ok := user.(models.UserModel)
+	if !ok {
+		error_handler.HandleError(c, http.StatusInternalServerError, "User not of type UserModel", []error{errors.New("user not of type UserModel")})
+		return
+	}
 	i := c.Param("id")
+
+	var owner string
+	err := s.NewDB.Get(&owner, `SELECT author FROM recipes WHERE id = $1`, i)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			error_handler.HandleError(c, http.StatusNotFound, "Recipe doesn't exist", []error{errors.New("recipe doesn't exist")})
+			return
+		}
+		if pqErr, ok := err.(*pq.Error); ok && pqErr.Code == "22P02" { // 22P02 is the code for invalid_text_representation
+			log.Printf("Potential SQL injection or invalid input detected, with ip: %s, and input %s", c.ClientIP(), i)
+			error_handler.HandleError(
+				c,
+				http.StatusBadRequest, fmt.Sprintf(`Value "%s" is not an ID`, i),
+				[]error{fmt.Errorf(`value "%s" is not an ID`, i)},
+			)
+			return
+		}
+		error_handler.HandleError(c, http.StatusInternalServerError, err.Error(), []error{err})
+		return
+	}
+	if owner != userModel.ID {
+		error_handler.HandleError(c, http.StatusUnauthorized, "User is not the owner of the recipe", []error{errors.New("user is not the owner of the recipe")})
+		return
+	}
 
 	result, err := s.NewDB.Exec(`DELETE FROM public.recipes WHERE id = $1`, i)
 	if err != nil {
@@ -222,7 +258,7 @@ func (s *Server) DeleteRecipe(c *gin.Context) {
 			error_handler.HandleError(
 				c,
 				http.StatusBadRequest, fmt.Sprintf(`Value "%s" is not an ID`, i),
-				[]error{errors.New(fmt.Sprintf(`Value "%s" is not an ID`, i))},
+				[]error{fmt.Errorf(`value "%s" is not an ID`, i)},
 			)
 			return
 		}
